Enable observables in plotting.plot_number_observer (#77)
* Session.run accepts numpy.ndarray as t

* Arithmetic representations in y

* Enable to set own observable in _plotly

* Split elegans-based plot functions into _elegans

* Format before assignment

* Rename plotting.default -> BACKEND

* Enable rule-based descriptions in observables
